# backend/cmd/api/docs/swagger.yaml
definitions:
...
  models.AddAddress:
    properties:
      city:
        type: string
      house_name:
        type: string
      name:
        type: string
      phone:
        type: string
      pin:
        type: string
      state:
        type: string
      street:
        type: string
    required:
    - city
    - house_name
    - name
    - pin
    - state
    - street
    type: object
...
  models.EditEmail:
    properties:
      email:
        type: string
    type: object
  models.EditName:
    properties:
      name:
        type: string
    type: object
  models.EditPhone:
    properties:
      phone:
        type: string
    type: object
...
  models.ForgotVerify:
    properties:
      newpassword:
        type: string
      otp:
        type: string
      phone:
        type: string
    type: object
...
  models.UserDetails:
    properties:
      confirmpassword:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
  models.UserLogin:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /users/cart:
    get:
      consumes:
      - application/json
      description: user can view their cart details
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get Cart
      tags:
      - User
  /users/cart/remove:
    delete:
      consumes:
      - application/json
      description: user can remove products from their cart
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Remove from Cart
      tags:
      - User
  /users/cart/updateQuantity/minus:
    put:
      consumes:
      - application/json
      description: user can subtract 1 quantity of product from their cart
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      - description: inv_id
        in: query
        name: inventory
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Subtract quantity in cart by one
      tags:
      - User
  /users/cart/updateQuantity/plus:
    put:
      consumes:
      - application/json
      description: user can add 1 quantity of product to their cart
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      - description: inv_id
        in: query
        name: inventory
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Add quantity in cart by one
      tags:
      - User
...
  /users/forgot-password:
    get:
      consumes:
      - application/json
      description: user can change their password if user forgot the password and
        login
      parameters:
      - description: forgot-verify
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/models.ForgotVerify'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Forgot password Verfy and Change
      tags:
      - User
...
  /users/login:
    post:
      consumes:
      - application/json
      description: user can log in by giving their details
      parameters:
      - description: login
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: User Login
      tags:
      - User
...
  /users/profile/address:
    get:
      consumes:
      - application/json
      description: user can get all their addresses
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get Addresses
      tags:
      - User
  /users/profile/address/add:
    post:
      consumes:
      - application/json
      description: user can add their addresses
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      - description: address
        in: body
        name: address
        required: true
        schema:
          $ref: '#/definitions/models.AddAddress'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Add Address
      tags:
      - User
  /users/profile/details:
    get:
      consumes:
      - application/json
      description: user can get all their details
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get User Details
      tags:
      - User
  /users/profile/edit/email:
    put:
      consumes:
      - application/json
      description: user can change their Email
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      - description: edit-email
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/models.EditEmail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Edit Email
      tags:
      - User
  /users/profile/edit/name:
    put:
      consumes:
      - application/json
      description: user can change their name
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      - description: edit-name
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/models.EditName'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Edit Name
      tags:
      - User
  /users/profile/edit/phone:
    put:
      consumes:
      - application/json
      description: user can change their Phone
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      - description: edit-phone
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/models.EditPhone'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Edit Phone
      tags:
      - User
  /users/profile/get-link:
    get:
      consumes:
      - application/json
      description: user can get a referral link and it is a share able link
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get Referral link
      tags:
      - User
...
  /users/profile/security/change-password:
    put:
      consumes:
      - application/json
      description: user can change their password
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Change Password
      tags:
      - User
...
  /users/signup:
    post:
      consumes:
      - application/json
      description: user can signup by giving their details
      parameters:
      - description: signup
        in: body
        name: signup
        required: true
        schema:
          $ref: '#/definitions/models.UserDetails'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: User Signup
      tags:
      - User
...
swagger: "2.0"